+++ examples/multi-agent-systems/room_mas/room_mas.py
@@ -12,9 +12,7 @@ def run_example(until, with_plots=True, log_level=logging.INFO):
     # Start by setting the log-level
     logging.basicConfig(level=log_level)
 
-    t_sample = 3600
-
-    env_config = {"rt": False, "t_sample": t_sample, "clock": True}
+    env_config = {"rt": False, "t_sample": 60, "clock": True}
 
     # Change the working directly so that relative paths work
     os.chdir(os.path.abspath(os.path.dirname(__file__)))